requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
numpy==1.26.0
python-dotenv==1.0.0
pandas==1.3.5